### 2020-06-18 #2355 @whedon
[PRE REVIEW]: set6: R6 Mathematical Sets Interface
**Submitting author:** @RaphaelS1 (<a href="http://orcid.org/0000-0001-9225-4654">Raphael E. B. Sonabend</a>)
**Repository:** <a href="https://github.com/xoopR/set6" target ="_blank">https://github.com/xoopR/set6</a>
**Version:** v0.1.4
**Reviewers:** @dmbates, davidjohannesmeyer, @davidjohannesmeyer
