+++ .github/workflows/main.yml
@@ -1,162 +1,0 @@
-on:
-  push:
-  pull_request:
-  # for now, CRON jobs only run on the default branch of the repo (i.e. usually on master)
-  schedule:
-    # * is a special character in YAML so you have to quote this string
-    - cron: "0 4 * * *"
-
-name: R CMD Check via {tic}
-
-jobs:
-  R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
-
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-    if: "!contains(github.event.head_commit.message, 'skip ci')"
-
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          # comment out lines if you do not want to build on certain platforms
-          - { os: windows-latest, r: "release" }
-          - { os: macOS-latest, r: "release" }
-   #       - { os: ubuntu-latest, r: "devel" }
-          - { os: ubuntu-latest, r: "release" }
-
-    env:
-      # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      CRAN: ${{ matrix.config.cran }}
-      # we are not allowed to write to ~/.ccache on GH Actions
-      # setting some ccache options
-      CCACHE_BASEDIR: ${{ GITHUB.WORKSPACE }}
-      CCACHE_DIR: ${{ GITHUB.WORKSPACE }}/.ccache
-      CCACHE_NOHASHDIR: true
-      CCACHE_SLOPPINESS: include_file_ctime
-      # make sure to run `tic::use_ghactions_deploy()` to set up deployment
-      TIC_DEPLOY_KEY: ${{ secrets.TIC_DEPLOY_KEY }}
-      # prevent rgl issues because no X11 display is available
-      RGL_USE_NULL: true
-
-    steps:
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-r@master
-        with:
-          r-version: ${{ matrix.config.r }}
-          Ncpus: 4
-
-      # LaTeX. Installation time:
-      # Linux: ~ 1 min
-      # macOS: ~ 1 min 30s
-      # Windows: never finishes
-      - uses: r-lib/actions/setup-tinytex@v1
-        if: runner.os != 'Windows'
-
-      - uses: r-lib/actions/setup-pandoc@master
-
-      # set date/week for use in cache creation
-      # https://github.community/t5/GitHub-Actions/How-to-set-and-access-a-Workflow-variable/m-p/42970
-      # - cache R packages daily
-      # - cache ccache weekly -> 'ccache' helps rebuilding the package cache faster
-      - name: "[Cache] Prepare daily timestamp for cache"
-        if: runner.os != 'Windows'
-        id: date
-        run: echo "::set-output name=date::$(date '+%d-%m')"
-
-      - name: "[Cache] Prepare weekly timestamp for cache"
-        if: runner.os != 'Windows'
-        id: datew
-        run: echo "::set-output name=datew::$(date '+%Y-%V')"
-
-      - name: "[Cache] Cache R packages"
-        if: runner.os != 'Windows'
-        uses: pat-s/always-upload-cache@v1.1.4
-        with:
-          path: ${{ env.R_LIBS_USER }}
-          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-          restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-
[…]
