+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: '2.2'
-pkgdown: 1.4.1.9000
-pkgdown_sha: 17d069073e39a73cf09b54eb756a83d55f8fa96b
+pandoc: 2.7.3
+pkgdown: 1.5.1
+pkgdown_sha: ~
 articles:
   set6: set6.html
-last_built: 2020-03-16T13:59Z
+last_built: 2020-06-19T12:34Z
 
