+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   set6: set6.html
-last_built: 2020-06-19T12:34Z
+last_built: 2020-06-20T04:18Z
 
