### 2020-06-22 bcae875
Deploy from Github Actions build 143150105 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/143150105
Commit: 3b1ad411a224b5108302c0a33b4008db1a94b05c
