### 2020-06-23 #2355 @arfon
Thanks @dpsanders - you can always grab the paper yourself with `@whedon assign me as editor`
