### 2020-06-24 #2355 @dpsanders
👋 Hi @RaphaelS1 -- thanks for submitting to JOSS!

Looking at the paper, I have a couple of questions and suggestions:

- Could you please add a couple of usage examples
- Describe in more detail how your package compares to the `sets` package
- What do you mean by the "symbolic representation" and "lazy evaluation" in this context? Maybe you could give examples.
- Please add the title, date and journal for the Meyer-Hornik paper in the bibliography.
- How does this compare to similar packages in other computer languages?

Thanks!
