### 2020-06-24 1bd97c0
paper updates
