+++ papers/JOSS/paper.bib
@@ -1,4 +1,3 @@
-```bib
 @misc{packageR,
 address = {Vienna},
 author = {{R Core Team}},
@@ -87,11 +86,11 @@ title = {{devtools: Tools to Make Developing R Packages Easier}},
 year = {2019}
 }
 
-@article{packagesets,
+@article{pkgsets,
 author = {Meyer, David and Hornik, Kurt},
-doi = {10.18637/jss.v031.i02}},
+doi = {10.18637/jss.v031.i02},
 journal = {Journal of Statistical Software},
-mendeley-groups = {Packages},
+mendeley-groups = {Software},
 number = {2},
 pages = {1----27},
 title = {{Generalized and Customizable Sets in R}},
@@ -109,6 +108,29 @@ url = {https://cran.r-project.org/package=distr6},
 year = {2019}
 }
 
+@article{pkgrcpp,
+author = {Eddelbuettel, Dirk and Francois, Romain},
+doi = {10.18637/jss.v040.i08},
+journal = {Journal of Statistical Software},
+number = {8},
+pages = {1--18},
+publisher = {CRAN},
+title = {{Rcpp: Seamless R and C++ Integration}},
+url = {http://www.jstatsoft.org/v40/i08/},
+volume = {40},
+year = {2011}
+}
 
-```
+@misc{pkgportion,
+author = {Decan, Alexandre},
+publisher = {PyPi},
+title = {portion - data structure and operations for intervals},
+url = {https://pypi.org/project/portion/},
+year = {2020}
+}
 
+@misc{pkgintervalsets,
+author = {Navelkar, Anchit},
+title = {{IntervalSets.jl}},
+year = {2016}
+}
