+++ papers/JOSS/paper.md
@@ -26,19 +26,71 @@ set6 makes use of the R6 object-oriented paradigm to introduce classes for impor
 
 The speed and efficiency of ``R6`` allows `set6` to be a scalable and efficient interface. A focus on symbolic representation and neat printing methods means that `set6` can accurately and clearly represent complicated sets. `set6` has the ambitious long-term goal of being the only dependency package needed for object-oriented interfaces in `R` that require clear symbolic representations of mathematical sets.
 
-Related software includes the `sets` [@packagesets] family of packages.
+Related software in `R` includes the `sets` [@packagesets] package, which uses the S3 and S4 object-oriented paradigms to implement mathematical sets. Whilst `sets` and `set6` have similar implemented features, as both offer multiple forms of sets (intervals, fuzzy, etc.) and with symbolic representation, the `sets` package does not offer lazy evaluation in set operations and thus can lead to significant overheads for large or possibly-infinite sets. Similar packages in other computing languages include i) `portion` [@pkgportion] in Python, which only supports intervals without generalisation to other mathematical sets; ii) `IntervalSets.jl` [@pkgintervalsets] in Julia, which is again limited to intervals though with good features for infix operators and inspection; and iii) a variety of implementations in the `JuliaIntervals` family of packages (https://juliapackages.com/u/juliaintervals), which primarily focuses on rigorous arithmetic implementation.  
+
+The example below demonstrates construction of a set and interval, comparisons of these, set complement operator, and printing of the final result.
+
+```R
+> a <- Set$new(1, 2, 3)
+> a$print()
+{1, 2, 3}
+> a$contains(1, "a")
+[1] TRUE FALSE
+> b <- Interval$new(1, Inf, class = "numeric")
+> b$isSubset(a)
+TRUE
+> c <- b - a
+> c$print()
+(1,2) ∪ (2,3) ∪ (3,+∞] 
+```
+
+
 
 # Key Design Principles
 
 1. **Maximum user-control over set operations** - Users can select how operations act on sets, including a choice of  associativity, lazy evaluation, and unicode printing.
 2. **Minimal dependencies** - `set6` has the goal of being a key dependency to any object-oriented requiring representation of mathematical sets, for example for representing function inputs and supports. Therefore `set6` is itself dependent on only three packages.
 3. **Inspectability and reactive user interface** - `set6` prioritises symbolic representation and lazy evaluation to allow for the package to be scalable and to fit into any other package. However it is ensured that this does not detract from a clear user interface, i.e. at all times it should be clear what an object contains both externally (how it prints) and internally (inspection methods). `set6` allows sets to be queried in many different ways, including calling the elements in the set (if finite), finding the bounds of the set (if numeric), and listing properties and traits.
+
+```R
+# Symbolic representation allows neat representation of infinite sets and intervals whilst making the elements clear by inspection
+
+> I <- Interval$new(10, Inf, type = "[)")
+> I$print()
+[10,+∞)
+> I$contains(9:11)
+[1] FALSE TRUE TRUE
+ 
+> N <- Naturals$new()
+> N$print()
+ℕ0 
+# binary operators also available
+> c(pi, 2) %inset% N
+[1] FALSE TRUE
+ 
+```
+
 4. **Combination of lazy and greedy evaluation** - By default, 'multiplying' operations such as products and powersets are evaluated lazily, whereas 'adding' operations such as unions and differences are evaluated greedily. These prevent system crashes from trying to evaluate sets of very large cardinality. In all cases, the user can override defaults. Symbolic representation is used with lazy evaluation so that large sets can be printed neatly without the individual elements ever being evaluated.
+
+```R
+# Lazy evaluation allows very large sets to be queried without explicit evaluation
+
[…]
