+++ papers/JOSS/paper.bib
@@ -131,6 +131,9 @@ year = {2020}
 
 @misc{pkgintervalsets,
 author = {Navelkar, Anchit},
+mendeley-groups = {Software},
 title = {{IntervalSets.jl}},
+url = {https://github.com/JuliaMath/IntervalSets.jl},
 year = {2016}
 }
+
