### 2020-06-24 c4bd849
Deploy from Github Actions build 145910582 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/145910582
Commit: 1bd97c00fcdb492d34ae2c30efedaf56bde65024
