### 2020-06-24 a533732
Deploy from Github Actions build 145915792 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/145915792
Commit: 3cc53b8125620684b7d97e40c50af5b3e1999152
