### 2020-06-24 #2355 @RaphaelS1
Hi @dpsanders thanks for the speedy feedback! I've added to the paper as suggested and pushed the latest copy to the same repo as before (I have assumed the questions were intended to be answered in the paper and not here but let me know otherwise).
