### 2020-06-25 #2355 @dpsanders
Thanks @RaphaelS1. You can make sure that the latest version is correct with the following command
