### 2020-06-26 d612fc4
Deploy from Github Actions build 148251613 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/148251613
Commit: 3379d2b50fe1446b06097bce0c9a7eba820a6c8c
