### 2020-06-27 #2355 @dpsanders
Thanks @RaphaelS1, the new version of the paper nicely explains the scope and use of the package. BTW, some more possibly relevant Julia packages include LazySets.jl, DomainSets.jl and SpecialSets.jl. But for example your lazy power set functionality looks pretty unique to me.
