### 2020-07-01 14b39fc
update lifecycle, version
