+++ NEWS.md
@@ -1,4 +1,4 @@
-# set6 0.1.4.9000
+# set6 0.1.5
 
 * Added `LogicalSet`, the set of $\{TRUE, FALSE\}$
 * Added `as.Set.numeric` and `as.Tuple.numeric`
