### 2020-07-02 #2355 @dpsanders
:wave: Hi @mcavs, @chlalanne and @drvinceknight! Would any of you be available and willing to review this submission to JOSS? Thanks!
