### 2020-07-02 243b818
lintr, op cleaner bugfix
