+++ NEWS.md
@@ -1,3 +1,7 @@
+# set6 0.1.5.9000
+
+* Bugfix in set operation cleaner
+
 # set6 0.1.5
 
 * Added `LogicalSet`, the set of $\{TRUE, FALSE\}$
