### 2020-07-04 #2355 @even4void
@dpsanders Unfortunately I don't really feel qualified to review this article. Maybe Dirk Eddelbuettel would be a better fit given the scope of the paper?
