### 2020-07-15 0032ec0
Deploy from Github Actions build 169552597 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/169552597
Commit: 243b81886247effe1ec3d8c04ff0a6a93a57ca5a
