### 2020-07-17 adb9f4d
interval contains bugfix
