+++ R/asFuzzySet.R
@@ -11,51 +11,57 @@
 #' * `as.FuzzySet.Set` - Creates a [FuzzySet] by assuming [Set] elements all have `membership` equal to \eqn{1}.
 #' * `as.FuzzySet.Interval` - First tries coercion via [as.Set.Interval] then uses [as.FuzzySet.Set].
 #' @export
-as.FuzzySet <- function(object){
-  UseMethod("as.FuzzySet",object)
+as.FuzzySet <- function(object) {
+  UseMethod("as.FuzzySet", object)
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.numeric <- function(object){
-  FuzzySet$new(elements = object[seq.int(1,length(object),2)],
-               membership = as.numeric(object[seq.int(2,length(object),2)]))
+as.FuzzySet.numeric <- function(object) {
+  FuzzySet$new(
+    elements = object[seq.int(1, length(object), 2)],
+    membership = as.numeric(object[seq.int(2, length(object), 2)])
+  )
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.list <- function(object){
+as.FuzzySet.list <- function(object) {
   return(FuzzySet$new(elements = object$elements, membership = object$membership))
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.matrix <- function(object){
-  return(FuzzySet$new(elements = object[,1], membership = object[,2]))
+as.FuzzySet.matrix <- function(object) {
+  return(FuzzySet$new(elements = object[, 1], membership = object[, 2]))
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.data.frame <- function(object){
-  if(all(c("elements", "membership") %in% colnames(object)))
+as.FuzzySet.data.frame <- function(object) {
+  if (all(c("elements", "membership") %in% colnames(object))) {
     return(FuzzySet$new(elements = object$elements, membership = object$membership))
-  else
-    return(FuzzySet$new(elements = object[,1], membership = object[,2]))
+  } else {
+    return(FuzzySet$new(elements = object[, 1], membership = object[, 2]))
+  }
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.Set <- function(object){
+as.FuzzySet.Set <- function(object) {
   return(FuzzySet$new(elements = object$elements))
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.FuzzySet <- function(object){
+as.FuzzySet.FuzzySet <- function(object) {
   return(FuzzySet$new(elements = object$elements, membership = object$membership()))
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.Interval <- function(object){
-  ifnerror(as.Set.Interval(object), stopwarn = "stop", errormsg = "Interval cannot be coerced to FuzzySet.")
+as.FuzzySet.Interval <- function(object) {
+  ifnerror(as.Set.Interval(object),
+    stopwarn = "stop",
+    errormsg = "Interval cannot be coerced to FuzzySet."
+  )
 }
 #' @rdname as.FuzzySet
 #' @export
-as.FuzzySet.ConditionalSet <- function(object){
+as.FuzzySet.ConditionalSet <- function(object) {
   stop("ConditionalSet cannot be coerced to FuzzySet.")
 }
 #-----------------------------
@@ -64,50 +70,53 @@ as.FuzzySet.ConditionalSet <- function(object){
 #' @rdname as.FuzzySet
 #' @aliases as.FuzzyTuple
 #' @export
-as.FuzzyTuple <- function(object){
-  UseMethod("as.FuzzyTuple",object)
[…]
