+++ cran-comments.md
@@ -1,3 +1,7 @@
+## Comments
+
+Early release due to urgent bugfix that is causing other package deployments to be delayed.
+
 ## Test Environments
 
 On GitHub actions:
