+++ R/atomic_coercions.R
@@ -1,12 +1,13 @@
 #' @export
-as.double.Set <- function(x,...) {
-  if(testFuzzy(x))
+as.double.Set <- function(x, ...) {
+  if (testFuzzy(x)) {
     return(unlist(x$support()))
-  else
+  } else {
     return(unlist(x$elements))
+  }
 }
 
 #' @export
-as.character.Set <- function(x, n = 1e10, ...){
+as.character.Set <- function(x, n = 1e10, ...) {
   x$strprint(n = n)
 }
