+++ tests/testthat/test-set6news.R
@@ -2,6 +2,6 @@ library(testthat)
 
 context("set6news")
 
-test_that("set6news",{
+test_that("set6news", {
   expect_silent(set6News())
 })
