### 2020-07-17 273e33d
bugfix: unionset unwrapping products
