+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Bugfix in set operation cleaner
 * Bugfix causing `Interval$contains` to return `TRUE` for tuples
+* Bugfix in union sets incorrectly unwrapping products
 
 # set6 0.1.5
 
