+++ R/operation_setpower.R
@@ -1,17 +1,20 @@
 #' @name setpower
 #' @rdname setpower
 #' @param x Set
-#' @param power power to raise set to
-#' @param simplify logical, if `TRUE` returns the result in its simplest (unwrapped) form, usually a `Set`,
-#' otherwise an `ExponentSet`.
+#' @param power power to raise set to, if `"n"` then a variable dimension set is created,
+#' see examples.`
+#' @param simplify logical, if `TRUE` returns the result in its simplest (unwrapped) form, usually
+#' a `Set`, otherwise an `ExponentSet`.
 #' @param nest logical, if `FALSE` (default) returns the n-ary cartesian product, otherwise returns
 #' the cartesian product applied n times.
 #' [Set]s. See details and examples.
 #' @title Power of a Set
 #' @return An R6 object of class `Set` or `ExponentSet` inheriting from `ProductSet`.
-#' @description A convenience wrapper for the n-ary cartesian product of a `Set` by itself, possibly multiple times.
-#' @details See the details of [setproduct] for a longer discussion on the use of the `nest` argument, in particular
-#' with regards to n-ary cartesian products vs. 'standard' cartesian products.
+#' @description A convenience wrapper for the n-ary cartesian product of a `Set` by itself,
+#' possibly multiple times.
+#' @details See the details of [setproduct] for a longer discussion on the use of the `nest`
+#' argument, in particular with regards to n-ary cartesian products vs. 'standard' cartesian
+#' products.
 #'
 #' @family operators
 #' @examples
@@ -32,17 +35,27 @@
 #'
 #' # Power of FuzzySet
 #' FuzzySet$new(1, 0.1, 2, 0.5)^2
+#'
+#' # Variable length
+#' x <- Interval$new(0, 1)^"n"
+#' x$contains(Tuple$new(0))
+#' x$contains(Tuple$new(0, 1))
+#' x$contains(Tuple$new(0, 1, 0, 0, 1, 1, 0))
+#' x$contains(Tuple$new(0, 2))
+#'
 #' @export
 setpower <- function(x, power, simplify = FALSE, nest = FALSE) {
   if (getR6Class(x) == "UniversalSet") {
     return(x)
   }
 
-  if (power == 1) {
+  if (power == 0) {
+    return(Set$new())
+  } else if (power == 1) {
     return(x)
-  }
-
-  if (getR6Class(x) %in% c("Set", "FuzzySet", "Tuple", "FuzzyTuple") & simplify) {
+  } else if (power == "n") {
+    return(ExponentSet$new(x, power))
+  } else if (getR6Class(x) %in% c("Set", "FuzzySet", "Tuple", "FuzzyTuple") & simplify) {
     x <- rep(list(x), power)
     return(do.call(setproduct, c(x, list(nest = nest, simplify = TRUE))))
   } else if (inherits(x, "ExponentSet")) {
