+++ cran-comments.md
@@ -1,6 +1,6 @@
 ## Comments
 
-Early release due to urgent bugfix that is causing other package deployments to be delayed.
+Sorry for second submission. Previous submission broke reverse dependencies in an unexpected way. Breakages now in `distr6` and `mlr3proba` are expected and patches are ready to fix them.
 
 ## Test Environments
 
@@ -16,4 +16,4 @@ On GitHub actions:
 There were no ERRORs, WARNINGs, or NOTEs.
 
 ## Downstream dependencies
-All downstream dependencies OK.
+ Breakages in `distr6` and `mlr3proba` are expected and patches are ready to fix them.
