### 2020-07-17 18bfac4
docs, fixes in variable exponent
