+++ tests/testthat/test_operations_setpower.R
@@ -37,4 +37,11 @@ test_that("variable", {
   expect_true(x$contains(Tuple$new(0, 1, 0, 0, 1, 1, 0)))
   expect_false(x$contains(Tuple$new(0, 2)))
   expect_error(x$contains(1))
+
+  expect_equal(x$contains(list(Tuple$new(1), Tuple$new(2))), c(TRUE, FALSE))
+  expect_equal(x$contains(list(Tuple$new(1, 0), Tuple$new(1, 1))), c(TRUE, TRUE))
+
+  expect_equal(x * Reals$new(), x)
+  expect_equal((Reals$new()^"n") * (Reals$new()^2), Reals$new()^"n")
+  expect_equal(x^2, x)
 })
