+++ man/useUnicode.Rd
@@ -20,7 +20,7 @@ unicode representations should be used, or standard alpha-numeric and special ch
 By default \code{set6} starts with unicode printing turned on.
 }
 \examples{
-current = useUnicode()
+current <- useUnicode()
 useUnicode(TRUE)
 useUnicode()
 useUnicode(current)

[…]
