+++ NEWS.md
@@ -1,4 +1,4 @@
-# set6 0.1.7
+# set6 0.1.7.9000
 
 * Critical patch
 
