### 2020-07-26 cce8ecd
Deploy from Github Actions build 182844269 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/182844269
Commit: 4250844870940c3c3f3c907b498f1cac52c5b91e
