+++ .github/workflows/tic.yml
@@ -22,8 +22,9 @@ jobs:
         config:
           # use a different tic template type if you do not want to build on all listed platforms
           - { os: windows-latest, r: "release" }
+          - { os: windows-latest, r: "devel" }
+          - { os: windows-latest, r: "3.6" }
           - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
-          - { os: macOS-latest, r: "devel" }
           - { os: ubuntu-latest, r: "release" }
 
     env:
