### 2020-07-26 28952da
R-devel patch
