+++ tests/testthat/test_operations_setunion.R
@@ -51,22 +51,24 @@ test_that("fuzzy", {
 })
 
 test_that("conditional", {
+  useUnicode(FALSE)
   expect_equal(
-    ConditionalSet$new(function(x) x == 0) + ConditionalSet$new(function(y) y > 0),
-    ConditionalSet$new(function(x, y) x == 0 | y > 0)
+    (ConditionalSet$new(function(x) x == 0) + ConditionalSet$new(function(y) y > 0))$strprint(),
+    "{x == 0 | y > 0 : x in V, y in V}"
   )
   expect_equal(
     setunion(
       ConditionalSet$new(function(x) x == 0), ConditionalSet$new(function(y) y > 0),
       ConditionalSet$new(function(z) z == 2)
-    ),
-    ConditionalSet$new(function(x, y, z) x == 0 | y > 0 | z == 2)
+    )$strprint(),
+    "{x == 0 | y > 0 | z == 2 : x in V, y in V, z in V}"
   )
   expect_equal(
-    ConditionalSet$new(function(x) x == 0) + ConditionalSet$new(function(y) y > 0) +
-      ConditionalSet$new(function(z) z == 2),
-    ConditionalSet$new(function(x, y, z) x == 0 | y > 0 | z == 2)
+    (ConditionalSet$new(function(x) x == 0) + ConditionalSet$new(function(y) y > 0) +
+      ConditionalSet$new(function(z) z == 2))$strprint(),
+    "{x == 0 | y > 0 | z == 2 : x in V, y in V, z in V}"
   )
+  useUnicode(TRUE)
 })
 
 test_that("contains", {
