### 2020-07-26 5a74439
remove news test, add nocov
