+++ tests/testthat/test-set6news.R
@@ -1,7 +1,0 @@
-library(testthat)
-
-context("set6news")
-
-test_that("set6news", {
-  expect_silent(set6News())
-})
