### 2020-07-26 cd3104c
update cran comments
