+++ cran-comments.md
@@ -1,14 +1,15 @@
 ## Comments
 
-Sorry for second submission. Previous submission broke reverse dependencies in an unexpected way. Breakages now in `distr6` and `mlr3proba` are expected and patches are ready to fix them.
+Early submission due to errors in R-devel.
 
 ## Test Environments
 
 On GitHub actions:
 
 * windows-latest (R release)
+* windows-latest (R devel)
+* windows-latest (R 3.6)
 * macOS-latest (R release)
-* macOS-latest (R devel)
 * ubuntu-latest (R release)
 
 ## R CMD check results
@@ -16,4 +17,4 @@ On GitHub actions:
 There were no ERRORs, WARNINGs, or NOTEs.
 
 ## Downstream dependencies
- Breakages in `distr6` and `mlr3proba` are expected and patches are ready to fix them.
+All downstream dependencies OK.
