### 2020-08-06 3d778cc
Deploy from Github Actions build 197186329 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/197186329
Commit: cd3104c69f568c06a269c29448266284a2351da1
