### 2020-08-24 600a0c0
Deploy from Github Actions build 221444924 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/221444924
Commit: cd3104c69f568c06a269c29448266284a2351da1
