### 2020-08-24 #2598 @whedon
[REVIEW]: set6: R6 Mathematical Sets Interface
**Submitting author:** @RaphaelS1 (<a href="http://orcid.org/0000-0001-9225-4654">Raphael E. B. Sonabend</a>)
**Repository:** <a href="https://github.com/xoopR/set6" target ="_blank">https://github.com/xoopR/set6</a>
**Version:** 0.2.0

----

----
