### 2020-08-26 b40f0fd
Deploy from Github Actions build 224752872 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/224752872
Commit: cd3104c69f568c06a269c29448266284a2351da1
