### 2020-08-27 cf393f8
Deploy from Github Actions build 226428419 [ci skip]
Build URL: https://github.com/xoopR/set6/actions/runs/226428419
Commit: cd3104c69f568c06a269c29448266284a2351da1
