+++ pkgdown.yml
@@ -1,7 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 1.5.1
+pkgdown: 1.6.0
 pkgdown_sha: ~
-articles:
-  set6: set6.html
-last_built: 2020-09-07T04:28Z
+articles: []
+last_built: 2020-09-08T04:28Z
 
