+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.7.3
 pkgdown: 1.6.0
 pkgdown_sha: ~
 articles: []
-last_built: 2020-09-08T04:28Z
+last_built: 2020-09-09T04:26Z
 
