### 2020-09-12 #2598 @dpsanders
👋 Hi @davidjohannesmeyer, @dmbates and @wch. Just checking in to see how your reviews are going.
Please feel free to open issues directly in the repo, https://github.com/xoopR/set6. Thanks!
