+++ pkgdown.yml
@@ -1,6 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 1.6.0
+pkgdown: 1.6.1
 pkgdown_sha: ~
-articles: []
-last_built: 2020-09-12T04:28Z
+articles:
+  set6: set6.html
+last_built: 2020-09-13T04:28Z
 
