+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   set6: set6.html
-last_built: 2020-09-13T04:28Z
+last_built: 2020-09-14T04:28Z
 
