+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   set6: set6.html
-last_built: 2020-09-17T04:29Z
+last_built: 2020-09-18T04:26Z
 
