### 2020-09-27 #2598 @dpsanders
👋 Hi @davidjohannesmeyer, @dmbates and @wch: I'm checking in to check on progress with your reviews for the `set6` submission to JOSS.
