### 2020-09-28 xoopR/set6#54 @dmbates
Unable to reproduce example in JOSS paper
I am reviewing https://github.com/openjournals/joss-reviews/ #2598 and I am unable to reproduce the first example in the paper
```R
> library(set6)

-----------------------------
	set6 v0.1.8

Get started:	?set6
Changelog:	set6News()
-----------------------------

> a <- Set$new(1,2,3)
> a
{1, 2, 3}
> a$contains(1)
[1] TRUE
> a$contains(pi)
[1] FALSE
> a$contains(1, pi, "a")
[1] TRUE
> a$contains(1, "a")
Error in if (all) { : argument is not interpretable as logical
```
Have I misunderstood the example?  (My R programming is a bit rusty so my apologies if I misunderstood)

If it matters this is on Ubuntu 20.04 with
Package: r-base-core
Version: 4.0.2-1.2004.0
