### 2020-09-28 #2598 @dmbates
I have opened https://github.com/xoopR/set6/issues/54 regarding one of the examples - the first one, actually.  I also see a couple of typos - e.g. under Key Design Principles, 2. Key Dependencies should read "... to any object-oriented package requiring ...".   @dpsanders where should I report a typo?
