### 2020-09-28 #2598 @dmbates
Could someone remind me how to run the package checks?  I cloned the git repository and tried a simple but unsuccessful approach
```
$ R CMD check set6
* using log directory ‘/home/bates/git/set6.Rcheck’
* using R version 4.0.2 (2020-06-22)
* using platform: x86_64-pc-linux-gnu (64-bit)
* using session charset: UTF-8
* checking for file ‘set6/DESCRIPTION’ ... ERROR
Required fields missing or empty:
  ‘Author’ ‘Maintainer’
* DONE

Status: 1 ERROR
See
  ‘/home/bates/git/set6.Rcheck/00check.log’
for details.
```
Do I need to run roxygen in some way on the directory?  (I know it is surprising that I am so inept with R these days but I have been working almost exclusively with Julia for several years now.)
