### 2020-09-28 #2598 @dpsanders
@dmbates You can e.g. make an issue on the repo with a list of typos, or even make Pull Requests. (That's easy to do using the pencil icon on the GitHub interface if they're minimal changes, for example.). Thanks!
