### 2020-09-28 #2598 @davidjohannesmeyer
You need to build it first, and check the .tar.gz file. Seems to pass.
