### 2020-09-28 xoopR/set6#54 @RaphaelS1
Sorry about this, this is a typo in the paper, it should read:

```R
a <- Set$new(1, 2, 3)
a$contains(c(1, "a"))
```

So in your examples it should be

```R
a <- Set$new(1,2,3)
a
a$contains(1)
a$contains(pi)
a$contains(c(1, pi, "a"))
a$contains(c(1, "a"))
```

or

```R
a$contains(list(1, pi, "a"))
a$contains(list(1, "a"))
```
