### 2020-09-28 #2598 @RaphaelS1
> I have opened [xoopR/set6#54](https://github.com/xoopR/set6/issues/54) regarding one of the examples - the first one, actually. I also see a couple of typos - e.g. under Key Design Principles, 2. Key Dependencies should read "... to any object-oriented package requiring ...". @dpsanders where should I report a typo?

Replying to the issue here so other reviewers can see. Apologies but there is indeed a typo in the first example, it reads

```R
> a <- Set$new(1, 2, 3) > a$print()
{1, 2, 3}
> a$contains(1, "a")
```

It should read

```R
> a <- Set$new(1, 2, 3) > a$print()
{1, 2, 3}
> a$contains(c(1, "a"))
```
