### 2020-09-29 xoopR/set6#55 @wch
JOSS review questions
This is for https://github.com/openjournals/joss-reviews/issues/2598.

Sorry to open a non-specific issue, but that seemed to make the most sense in this case (these individual questions didn't seem to deserve their own issues).

******

Is there a plan to implement `$isSubset` for union sets? Borrowing from one of the examples:

```R
a <- Set$new(1, 2, 3)
b <- Interval$new(1, Inf)
c <- b - a
c$print()
#> (1,2) ∪ (2,3) ∪ (3,+∞]

c$isSubset(Set$new(1))
#> isSubset currently not implemented for this wrapper.
```

******

Also, some small typos in the paper. It has:

```
> c$print()
(1,2) U (2,3) U (3,+Inf]
```

But it the output is now:

```
(1,2) ∪ (2,3) ∪ (3,+∞]
```

It looks like the union symbols and infinity symbol are different from the current code compared to what's in the paper.

Update: I just noticed that the output of other code examples has similar inconsistencies -- the output when I run it has special mathematical symbols, but the output in the paper does not.
