### 2020-09-30 #2598 @wch
I've finished my review. Other than the code error in the example and a few small comments I made in https://github.com/xoopR/set6/issues/55, it looks good. (If there's more needed from me for a review, please let me know -- the JOSS review process is unfamiliar to me.)
