### 2020-09-30 xoopR/set6#55 @RaphaelS1
Hi Winston,

Thanks for your review.

> Update: I just noticed that the output of other code examples has similar inconsistencies -- the output when I run it has special mathematical symbols, but the output in the paper does not.

The last paragraph of the summary in the paper: "The example below demonstrates construction of a set and interval, comparisons of these, the set complement operator, and printing of the final result. Note the example does not use unicode printing but that this is possible within the package."

Unicode printing is set to `l10n_info()$`UTF-8` by default. To reproduce the same printing as in the example just first run `useUnicode(TRUE)`. I can't remember if I choose not to use these symbols in the paper because of a problem with the markdown or just to have a consistent interface for uses of all platforms.

> Is there a plan to implement $isSubset for union sets?

Yes there are a few methods missing from wrappers, in most cases this is because I haven't worked out the most efficient implementation yet. Though in your specific example `c$contains(1)` is equivalent and `$contains` is implemented for all wrappers; therefore I'll most likely just implement it in a similar way, i.e. calling `$contains` on the set elements, but unsure for intervals how this would work at the moment
