+++ R/setSymbol.R
@@ -18,7 +18,8 @@ setSymbol <- function(set, zero = FALSE) {
       posreals = paste0("\u211D", zero_str, "+"),
       negreals = paste0("\u211D", zero_str, "-"),
       extendedreals = "\u211D \u222A {-\u221E, +\u221E}",
-      complex = "\u2102"
+      complex = "\u2102",
+      primes = "\u2118"
     ))
   } else {
     return(switch(set,
@@ -34,7 +35,8 @@ setSymbol <- function(set, zero = FALSE) {
       posreals = paste0("R", zero_str, "+"),
       negreals = paste0("R", zero_str, "-"),
       extendedreals = "R U {-Inf, +Inf}",
-      complex = "C"
+      complex = "C",
+      primes = "P"
     ))
   }
 }
