### 2020-10-13 xoopR/set6#55 @RaphaelS1
That would be up to @wch , if he's happy we can close this issue
