### 2020-10-17 #2598 @davidjohannesmeyer
Sorry for being so late. Here go my comments:

- Currently, the package can **not** handle sets of general R objects, because the equals() and contains() methods rely on the character representation of the elements. But as.character() does not "serialize" R objects - e.g., attributes are stripped, references are are treated all the same (try, e.g., sets of R6-classes) - and whenever specific as.character methods are implemented, the approach will fail. Better use identical() on the actual elements, instead of %in% on the character representation.
- The documentation claims that set6 aims to be an "upgrade" to the sets package. Currently, this is not the case - some features are extended (intervals), others are currently not implemented at all (e.g., multisets, fuzzy logic).
- The role of wrappers is unclear from the documentation.
- The cardinality is currently implemented as property. As this is not trivial to compute, especially for conditional sets or in case of lazy evaluation, I wonder whether this shouldn't become a method.
- Would it make sense to have a global option for the default Universal Set?
- Complex intervals are unclear to me, as complex numbers are unordered. At least, there should be some sort of definition in the help page.
- The representation of conditional sets seems unusual to me, as the universe is printed _after_ the condition. Isn't that usually written the other way round, e.g.: {x element R | x > 0}?
- Logicals are not mentioned as special set by listSpecialSets()
