### 2020-10-19 #2598 @RaphaelS1
Thanks for the detailed look at the package and the comments @davidjohannesmeyer.

> Currently, the package can not handle sets of general R objects, because the equals() and contains() methods rely on the character representation of the elements. But as.character() does not "serialize" R objects - e.g., attributes are stripped, references are are treated all the same (try, e.g., sets of R6-classes) - and whenever specific as.character methods are implemented, the approach will fail. Better use identical() on the actual elements, instead of %in% on the character representation.

Originally the objects were directly compared and not the character representations. `identical` doesn’t work for R6 objects (e.g. `identical(Set$new(1), Set$new(1))` returns `FALSE`), the alternative `all.equal` has significant overheads due to recursive unpacking of objects and their inherited classes (something I’ve/others previously discussed with @wch regarding a separate package). For now we will change the documentation to say that sets can handle any atomic or set6 object, which is certainly the case, and then work on a fix to handle all cases

> The documentation claims that set6 aims to be an "upgrade" to the sets package. Currently, this is not the case - some features are extended (intervals), others are currently not implemented at all (e.g., multisets, fuzzy logic).

Firstly to be clear, {set6} does not *currently* claim to be an upgrade to {sets} but instead sets this as a high bar to*aim* torwards, and secondly upgrade here refers more to S4 —> R6. With respect to comparison of implemented features, we can easily add multisets by copying and slightly modifying the tuple class, and fuzzy sets are implemented with fuzzy logic being an open issue in the repo. If there are any more key features you think are missing please feel free to tell me here or in an issue.

> The role of wrappers is unclear from the documentation.

Thanks we’ll improve the documentation. Their primary purpose is to fulfil the lazy evaluation of composite sets by overloading the representation and membership methods.

> The cardinality is currently implemented as property. As this is not trivial to compute, especially for conditional sets or in case of lazy evaluation, I wonder whether this shouldn't become a method.

 We can see good arguments for this as a method or property. However by keeping it as a property it clearly aids in the inspectability of an object, i.e. by calling `$properties` all necessary and useful results are returned — of course instead of storing `Property` objects we could instead have `$properties` call the methods, though possibly at a significant cost to runtime.

> Would it make sense to have a global option for the default Universal Set?

We have found that default arguments in functions are ‘easier’ for users than global options and moreover there may be some use-cases where the universe will be changed frequently.

> Complex intervals are unclear to me, as complex numbers are unordered. At least, there should be some sort of definition in the help page.

Agreed. Several of the SpecialSets are primarily for the `$contains` method and other methods will give strange results (e.g. rationals, primes, complex), this should be documented for all of them. To fix this I'll change the inheritance from `SpecialSet -> Interval` to `SpecialSet -> Set` and then only implement the methods that make sense, with clear documentation showing which do/don't exist for that special set and why.

> The representation of conditional sets seems unusual to me, as the universe is printed after the condition. Isn't that usually written the other way round, e.g.: {x element R | x > 0}?

Agreed this was a strange choice and will be fixed.

> Logicals are not mentioned as special set by listSpecialSets()

Thanks, this is a bug to be fixed.
