### 2020-10-22 a954516
Update tic.R
