### 2020-10-22 64cee9c
Update tic.R
