### 2020-10-22 xoopR/set6#56 @RaphaelS1
Add Logicals and Universal to listSpecialSets
See https://github.com/openjournals/joss-reviews/issues/2598
