### 2020-10-22 xoopR/set6#57 @RaphaelS1
Reverse string representation of conditional sets
See https://github.com/openjournals/joss-reviews/issues/2598
