### 2020-10-22 xoopR/set6#58 @RaphaelS1
Fix inheritance of SpecialSets
From SpecialSets -> Interval to SpecialSets -> Set; then just inherit each method as required. Alternatively could remove the SpecialSet class altogether and instead just have `listSpecialSets` and manually add to this, also `family` tag in roxygen. Then special intervals can inherit from `Interval` or otherwise from `Set`, this is probably the cleanest way.

See https://github.com/openjournals/joss-reviews/issues/2598
