### 2020-10-22 xoopR/set6#59 @RaphaelS1
Improve documentation for Wrappers
See https://github.com/openjournals/joss-reviews/issues/2598
