### 2020-10-22 xoopR/set6#60 @RaphaelS1
Add multisets
Almost identical to Tuples, should be simple copy/paste with minor edits.

See https://github.com/openjournals/joss-reviews/issues/2598

* [x] Multiset class
* [x] FuzzyMultiset class
* [x] Multiset assertions
* [x] FuzzyMultiset assertions
* [x] Multiset coercions to/from
* [x] FuzzyMultiset coercions to/from
* [x] Multiset operations
* [x] FuzzyMultiset operations
* [x] NEWS
