### 2020-10-22 xoopR/set6#61 @RaphaelS1
Fix Set$contains method/documentation
From https://github.com/openjournals/joss-reviews/issues/2598

> Currently, the package can not handle sets of general R objects, because the equals() and contains() methods rely on the character representation of the elements. But as.character() does not "serialize" R objects - e.g., attributes are stripped, references are are treated all the same (try, e.g., sets of R6-classes) - and whenever specific as.character methods are implemented, the approach will fail. Better use identical() on the actual elements, instead of %in% on the character representation.

`identical` doesn't work for R6, can consider `all.equal` but significant overheads. Currently comparisons will fail if there is no `as.character` method. Change documentation to state any atomic or set6 class can be handled, and in the future all will be. (`all.equal` may be forced)
