### 2020-10-22 b894329
Update tic.R
