+++ tic.R
@@ -11,7 +11,6 @@ get_stage("before_deploy") %>%
   do_pkgdown()
   
   get_stage("deploy") %>%
-  #  add_code_step(devtools::install()) %>%
     add_code_step(rmarkdown::render("README.Rmd")) %>%
     add_step(step_do_push_deploy(commit_paths = c("README.md")))
 }
