+++ R/Set_Interval_SpecialSet.R
@@ -20,7 +20,7 @@ SpecialSet <- R6Class("SpecialSet",
         stop(paste(getR6Class(self, pos = environment()), "is an abstract class that can't be initialized."))
       }
 
-      super$initialize(lower = lower, upper = upper, type = type, class = class, universe = UniversalSet$new())
+      super$initialize(lower = lower, upper = upper, type = type, class = class, universe = Universal$new())
 
       invisible(self)
     },
@@ -285,56 +285,6 @@ ExtendedReals <- R6Class("ExtendedReals",
     #' @return A new `ExtendedReals` object.
     initialize = function() {
       super$initialize(type = "[]")
-    }
-  )
-)
-
-#' @name Complex
-#' @title Set of Complex Numbers
-#' @description The mathematical set of complex numbers,
-#' defined as the the set of reals with possibly imaginary components. i.e.
-#' \deqn{\\{a + bi \\ : \\ a,b \in R\\}}{{a + bi : a,b \epsilon R}}
-#' where \eqn{R} is the set of reals.
-#' @details Unlike the other `SpecialSet`s, `Complex` can be used to define an `Interval`. In this
-#' case where values can be complex, as opposed to reals or integers in [Interval].
-#' @family special sets
-#' @export
-Complex <- R6Class("Complex",
-  inherit = SpecialSet,
-  public = list(
-    #' @description Create a new `Complex` object.
-    #' @return A new `Complex` object.
-    #' @param lower complex. Where to start the set.
-    #' @param upper complex. Where to end the set.
-    initialize = function(lower = -Inf + 0i, upper = Inf + 0i) {
-      super$initialize(lower = as.complex(lower), upper = as.complex(upper), type = "()", class = "numeric")
-    },
-
-    #' @description Tests to see if \code{x} is contained in the Set.
-    #'
-    #' @param x any. Object or vector of objects to test.
-    #' @param all logical. If `FALSE` tests each `x` separately. Otherwise returns `TRUE` only if all `x` pass test.
-    #' @param bound logical.
-    #'
-    #' @details \code{x} can be of any type, including a Set itself. \code{x} should be a tuple if
-    #' checking to see if it lies within a set of dimension greater than one. To test for multiple \code{x}
-    #' at the same time, then provide these as a list.
-    #'
-    #' If `all = TRUE` then returns `TRUE` if all `x` are contained in the `Set`, otherwise
-    #' returns a vector of logicals. For [Interval]s, `bound` is used to specify if elements lying on the
-    #' (possibly open) boundary of the interval are considered contained (`bound = TRUE`) or not (`bound = FALSE`).
-    #'
-    #' @return If \code{all} is `TRUE` then returns `TRUE` if all elements of \code{x} are contained in the `Set`, otherwise
-    #' `FALSE.` If \code{all} is `FALSE` then returns a vector of logicals corresponding to each individual
-    #' element of \code{x}.
-    #'
-    #' The infix operator `%inset%` is available to test if `x` is an element in the `Set`,
-    #' see examples.
-    contains = function(x, all = FALSE, bound = NULL) {
-      returner(
-        x = sapply(x, is.complex),
-        all = all
-      )
     }
   )
 )
