+++ tests/testthat/test_set_conditionalset.R
@@ -46,9 +46,9 @@ test_that("equals", {
 
 test_that("strprint", {
   useUnicode(TRUE)
-  expect_equal(ConditionalSet$new(function(x) TRUE)$strprint(), paste0("{TRUE : x", " \u2208 V}"))
+  expect_equal(ConditionalSet$new(function(x) TRUE)$strprint(), "{x \u2208 \U1D54D : TRUE}")
   useUnicode(FALSE)
-  expect_equal(ConditionalSet$new(function(x) TRUE)$strprint(), paste0("{TRUE : x in V}"))
+  expect_equal(ConditionalSet$new(function(x) TRUE)$strprint(), "{x in V : TRUE}")
   useUnicode(TRUE)
 })
 
