### 2020-10-22 bfe1437
WIP add multiset and fuzzymultiset
