+++ R/Set_Logicals.R
@@ -17,6 +17,7 @@ Logicals <- R6::R6Class("Logicals",
     initialize = function() {
       private$.elements <- list(TRUE, FALSE)
       private$.str_elements <- c("TRUE", "FALSE")
+      private$.multiplicity <- list("TRUE" = 1, "FALSE" = 1)
       private$.class <- "logical"
       private$.properties <- Properties$new(closure = "closed", cardinality = 2)
       private$.lower <- TRUE
