### 2020-10-22 9928ae4
fix conditionalset tests
