+++ tests/testthat/test_operations_setproduct.R
@@ -1,7 +1,3 @@
-library(testthat)
-
-context("setproduct")
-
 test_that("redundant", {
   expect_equal(setproduct(), Set$new())
   expect_equal(setproduct(Set$new(1)), Set$new(1))
@@ -35,7 +31,7 @@ test_that("Interval", {
   expect_equal(i$contains(list(Tuple$new(1, 4), Tuple$new(1, 3), Tuple$new(3, 1))), c(TRUE, TRUE, FALSE))
   expect_equal(
     (Interval$new(1, 2) * ConditionalSet$new(function(x) TRUE))$strprint(),
-    "[1,2] X {TRUE : x in V}"
+    "[1,2] X {x in V : TRUE}"
   )
   expect_equal(Interval$new(1, 2) * Set$new(), Interval$new(1, 2))
   useUnicode(TRUE)
@@ -45,7 +41,7 @@ test_that("conditionalset", {
   useUnicode(FALSE)
   expect_equal(
     (ConditionalSet$new(function(x) x == 1) * ConditionalSet$new(function(y) y > 1))$strprint(),
-    "{x == 1 : x in V} X {y > 1 : y in V}"
+    "{x in V : x == 1} X {y in V : y > 1}"
   )
   useUnicode(TRUE)
 })
@@ -57,7 +53,7 @@ test_that("fuzzy", {
   expect_equal(FuzzySet$new() * FuzzySet$new(1, 0.2), FuzzySet$new(1, 0.2))
   expect_equal(
     (FuzzySet$new(1, 0.5) * ConditionalSet$new(function(x) TRUE))$strprint(),
-    "{1(0.5)} X {TRUE : x in V}"
+    "{1(0.5)} X {x in V : TRUE}"
   )
   expect_true(setproduct(FuzzySet$new(1, 0.1), FuzzySet$new(2, 0.2), simplify = TRUE)$equals(Set$new(FuzzyTuple$new(1, 0.1, 2, 0.2))))
   expect_equal(
