### 2020-10-22 21e226a
Update set6.rmd
