+++ vignettes/set6.rmd
@@ -23,6 +23,7 @@ knitr::opts_chunk$set(collapse = T, comment = "#>")
 options(tibble.print_min = 4L, tibble.print_max = 4L)
 library("set6")
 set.seed(42)
+useUnicode(TRUE)
 ```
 
 ## Constructing a Set
@@ -38,7 +39,8 @@ empty$properties
 empty$traits
 ```
 
-Sets can take elements of any class, including other sets.
+Sets can contain any object in R that has a valid `as.character` dispatch method that ensures uniqueness in objects. This is a requirement as internally string representations are used to compare sets.
+
 ```{r}
 Set$new(1,2,3)
 Set$new(letters[1:5])
@@ -193,8 +195,7 @@ Set$new(1) & Set$new(2)
 Finally we look briefly at wrappers, and the `simplify` argument. 
 Each operation has an associated wrapper that will be created if `simplify == FALSE` or if the
 resulting set is too complicated to return as a single `Set` object. The operations concerned with
-products, i.e. `setproduct`, `setpower`, `powerset`, all have `simplify == FALSE` as the default; whereas
-the others concerned with unions and differences, have `simplify == TRUE` as the default.
+products, i.e. `setproduct`, `setpower`, `powerset`, all have `simplify == FALSE` as the default; whereas the others concerned with unions and differences, have `simplify == TRUE` as the default.
 
 ```{r}
 # default: simplify = TRUE
