### 2020-10-22 0bc6f85
Update NEWS.md
