+++ NEWS.md
@@ -3,6 +3,7 @@
 * UniversalSet renamed Universal, old class will be removed in v0.4.0.
 * LogicalSet renamed Logicals, old class will be removed in v0.4.0.
 * `Complex` now inherits from `Set`, incorrect methods for `isSubset, equals` have been removed.
+* Printing of `ConditionalSet` is fixed
 
 # set6 0.1.8
 
