+++ tests/testthat/test_set_logicals.R
@@ -1,10 +1,9 @@
 test_that("construction", {
   expect_silent(Logicals$new())
   l <- Logicals$new()
-  s <- Set$new(TRUE, FALSE)
-  expect_equal(l$elements, s$elements)
-  expect_equal(l$min, s$min)
-  expect_equal(l$max, s$max)
-  expect_equal(l$lower, s$lower)
-  expect_equal(l$upper, s$upper)
+  expect_equal(l$elements, list(TRUE, FALSE))
+  expect_equal(l$min, NaN)
+  expect_equal(l$max, NaN)
+  expect_equal(l$lower, TRUE)
+  expect_equal(l$upper, FALSE)
 })
