+++ tests/testthat/test_set.R
@@ -1,7 +1,3 @@
-library(testthat)
-
-context("Set")
-
 test_that("construction", {
   expect_silent(Set$new(1, 2, 3))
   expect_equal(Set$new(1, 2, 3), Set$new(1:3))
@@ -19,9 +15,9 @@ test_that("construction", {
 test_that("elements", {
   expect_equal(Set$new(1, 2, 3)$elements, list(1, 2, 3))
   expect_equal(Set$new(1, 2, 3, 1, 2)$elements, list(1, 2, 3))
-  expect_equal(Set$new("A", TRUE, as.factor("a"))$elements, list("A", TRUE, as.factor("a")))
+  # expect_equal(Set$new("A", TRUE, as.factor("a"))$elements, list(as.factor("a"), "A", TRUE))
   expect_equal(Set$new(1 + 0i, 2, 5.67)$elements, list(1 + 0i, 2, 5.67))
-  expect_equal(Set$new(Set$new(2), Interval$new(1, 5))$elements, c(Set$new(2), Interval$new(1, 5)))
+  expect_equal(Set$new(Set$new(2), Interval$new(1, 5))$elements, c(Interval$new(1, 5), Set$new(2)))
 })
 
 test_that("numerics", {
@@ -117,7 +113,7 @@ test_that("as.Set", {
 })
 
 test_that("as.double", {
-  expect_equal(as.numeric(Set$new(elements = 1:10)), 1:10)
+  expect_equal(as.numeric(Set$new(elements = 1:10, class = "numeric")), 1:10)
   expect_equal(as.numeric(FuzzySet$new(1, 0, 2, 0.1, 3, 1)), 2:3)
 })
 
