### 2020-10-23 180e39e
fix equals method
