+++ R/Set_FuzzySet_FuzzyMultiset.R
@@ -66,8 +66,8 @@ FuzzyMultiset <- R6Class("FuzzyMultiset",
           return(FALSE)
         }
 
-        ifelse(all.equal(el$multiplicity(), self$multiplicity()) == TRUE, TRUE, FALSE) &&
-          ifelse(all.equal(el$membership(), self$membership()) == TRUE, TRUE, FALSE)
+        ifelse(all(all.equal(el$multiplicity(), self$multiplicity()) == TRUE), TRUE, FALSE) &&
+          ifelse(all(all.equal(el$membership(), self$membership()) == TRUE), TRUE, FALSE)
       })
 
       returner(ret, all)
