+++ R/operation_setcomplement.R
@@ -54,12 +54,8 @@
 #' @export
 setcomplement <- function(x, y, simplify = TRUE) {
   if (missing(y)) {
-    if (getR6Class(x) == "FuzzySet") {
-      return(FuzzySet$new(elements = x$elements, membership = 1 - unlist(x$membership())))
-    } else if (getR6Class(x) == "FuzzyTuple") {
-      return(FuzzyTuple$new(elements = x$elements, membership = 1 - unlist(x$membership())))
-    } else if (getR6Class(x) == "FuzzyMultiset") {
-      return(FuzzyMultiset$new(elements = x$elements, membership = 1 - unlist(x$membership())))
+    if (testFuzzy(x)) {
+      return(getR6Class(x, FALSE)$new(elements = x$elements, membership = 1 - unlist(x$membership())))
     } else if (is.null(x$universe)) {
       stop("Set y is missing and x does not have a universe for absolute complement.")
     } else {
@@ -86,6 +82,8 @@ setcomplement <- function(x, y, simplify = TRUE) {
   if (testCountablyFinite(x) & testCountablyFinite(y)) {
     if (testTuple(x)) {
       return(Tuple$new(elements = setdiff(x$elements, y$elements)))
+    } else if (testMultiset(x)) {
+      return(Multiset$new(elements = setdiff(x$elements, y$elements)))
     } else {
       return(Set$new(elements = setdiff(x$elements, y$elements)))
     }
@@ -97,6 +95,8 @@ setcomplement <- function(x, y, simplify = TRUE) {
 
   if (testCrisp(x) & testFuzzyTuple(y)) {
     y <- as.Tuple(y)
+  } else if (testCrisp(x) & testFuzzyMultiset(y)) {
+    y <- as.Multiset(y)
   } else if (testCrisp(x) & testFuzzySet(y)) {
     y <- as.Set(y)
   }
@@ -127,18 +127,14 @@ setcomplement <- function(x, y, simplify = TRUE) {
 #' @export
 setcomplement.Set <- function(x, y, simplify = TRUE) {
   # difference of two sets
-  if (getR6Class(y) %in% c("Set", "Tuple", "FuzzySet", "FuzzyTuple")) {
-    if (testTuple(x)) {
-      return(Tuple$new(elements = x$elements[!(x$elements %in% y$elements)]))
-    } else {
-      return(Set$new(elements = x$elements[!(x$elements %in% y$elements)]))
+  if (getR6Class(y) %in% c("Set", "Tuple", "FuzzySet", "FuzzyTuple", "Multiset", "FuzzyMultiset")) {
+    if (getR6Class(x) %in% c("Set", "Tuple", "Multiset")) {
+      return(getR6Class(x, FALSE)$new(elements = x$elements[!(x$elements %in% y$elements)]))
     }
-    # difference of set and interval
+    # difference of set and interval - signif performance difference when separated from above
   } else if (testInterval(y)) {
-    if (testTuple(x)) {
-      return(Tuple$new(elements = x$elements[!y$contains(x$elements)]))
-    } else {
-      return(Set$new(elements = x$elements[!y$contains(x$elements)]))
+    if (getR6Class(x) %in% c("Set", "Tuple", "Multiset")) {
+      return(getR6Class(x, FALSE)$new(elements = x$elements[!y$contains(x$elements)]))
     }
   }
 }
@@ -224,16 +220,9 @@ setcomplement.Interval <- function(x, y, simplify = TRUE) {
 #' @rdname setcomplement
 #' @export
 setcomplement.FuzzySet <- function(x, y, simplify = TRUE) {
-  y <- as.FuzzySet(y)
[…]
