+++ NEWS.md
@@ -4,6 +4,7 @@
 * LogicalSet renamed Logicals, old class will be removed in v0.4.0.
 * `Complex` now inherits from `Set`, incorrect methods for `isSubset, equals` have been removed.
 * Printing of `ConditionalSet` is fixed
+* `Rationals` and child-classes now error on calls to `contains, isSubset, equals` as any prior results were likely wrong/misleading
 
 # set6 0.1.8
 
