### 2020-10-26 89b8db2
Improve wrapper documentation
Closes #59
