+++ README.Rmd
@@ -145,9 +145,8 @@ Whilst the `set6` API is stable, it is considered 'maturing', and therefore whil
 
 * Finalising all methods and fields - some are missing or possibly inaccurate for some wrappers. For example the cardinality of `ComplementSet`s is imprecise at the moment.
 * We are considering adding a `simplify` method to wrappers to reduce classes inheriting from `SetWrapper` to simpler sets. This allows users to perform operations with `simplify = FALSE` and then to change their mind.
-* There are some inefficient loops implemented that will need refactoring for more efficient code.
+* There are known bottlenecks that need to be fixed to massively improve speed and efficiency.
 * Adding more tutorials to make the interface easier for beginners, especially people new to R6
-
 
 At a later stage we may consider adding Venn diagrams for visualisation of sets and intervals, but this is very low priority.
 
