### 2020-10-26 #2598 @davidjohannesmeyer
Yes!

Best, David

> Am 26.10.2020 um 03:52 schrieb David P. Sanders <notifications@github.com>:
>
> ﻿
> @davidjohannesmeyer Thanks for your review. Are you happy with the changes?
>
