### 2020-10-29 xoopR/set6#64 @dpsanders
Related Julia software
The following related Julia software should perhaps be referenced in the paper:

- https://github.com/JuliaReach/LazySets.jl
- https://github.com/JuliaApproximation/DomainSets.jl

Part of JOSS review https://github.com/openjournals/joss-reviews/issues/2598
