### 2020-10-29 xoopR/set6#64 @RaphaelS1
Thanks David. I'm unsure about adding these as I'd argue that (just from briefly scanning the docs) both implement more complex mathematical objects that will likely never be in set6. As there are several Julia packages for more complicated handling of sets it could be clearer to instead just add a line pointing to Julia and the orgs instead?
