### 2020-11-01 xoopR/set6#64 @dpsanders
Sorry, for some reason I missed your comment.

Yes I agree they are more complex; I just felt that you implied in the paper that there were no Julia packages with lazy representations of sets. Maybe you can rewrite that paragraph. Thanks!
