+++ papers/JOSS/paper.bib
@@ -131,9 +131,24 @@ year = {2020}
 
 @misc{pkgintervalsets,
 author = {Navelkar, Anchit},
-mendeley-groups = {Software},
 title = {{IntervalSets.jl}},
 url = {https://github.com/JuliaMath/IntervalSets.jl},
 year = {2016}
+}
+
+
+@misc{pkglazysets,
+author = {Forets, Marcelo and Schilling, Christian},
+title = {{LazySets.jl}},
+url = {https://github.com/JuliaReach/LazySets.jl},
+year = {2018}
+}
+
+
+@misc{pkgdomainsets,
+author = {Huybrechs, Daan},
+title = {{DomainSets.jl}},
+url = {https://github.com/JuliaApproximation/DomainSets.jl},
+year = {2018}
 }
 
