### 2020-11-01 xoopR/set6#64 @RaphaelS1
Ah I've re-read the part you're referring to and see what you mean. I've added the software you mentioned above. Please close this issue if you're happy with this.
