### 2020-11-05 xoopR/set6#64 @RaphaelS1
Para 4:

" Similar packages in other computing languages include: i) portion (Decan, 2020) in Python, which only supports
intervals without generalisation to other mathematical sets; ii) IntervalSets.jl (Navelkar,
2016) in Julia, which is again limited to intervals though with good features for infix operators
and inspection; iii) a variety of packages in the JuliaIntervals suite (https://juliapackages.
com/u/juliaintervals), which primarily focus on rigorous arithmetic implementation; **and iv)
LazySets.jl (Forets & Schilling, 2018) and DomainSets.jl (Huybrechs, 2018) which both
faciliate symbolic representation of sets and infinite intervals in Julia.**"
