### 2020-11-05 #2598 @dpsanders
👋 @RaphaelS1: I think we're ready to proceed to publication!

Could you please archive the software e.g. at Zenodo, making sure that the title and authors are exactly as in the paper, then report the version number and DOI here? Thanks!
