### 2020-11-05 #2598 @RaphaelS1
@dpsanders Fantastic!

https://zenodo.org/record/4247626#.X6Q5KVP7TOQ
set6 0.2.0
http://doi.org/10.5281/zenodo.4247626
