+++ README.Rmd
@@ -157,7 +157,7 @@ tuple, intervals, and fuzzy variants. However the symbolic representation is not
 the package, does not allow for clear inspection of set/interval elements, and there is no support for
 multi-dimensional sets.
 
-* [BaseSet](https://github.com/llrs/BaseSet) - The **BaseSet** package focuses on storing and analysing
+* [BaseSet](https://github.com/ropensci/BaseSet) - The **BaseSet** package focuses on storing and analysing
 sets in a 'tidy' way, with more options for data storage in long and wide formats. The primary usage is
 neat and efficient inspection of finite sets, there is no support for infinite sets or symbolic
 representation.
