### 2020-11-09 #2598 @kthyng
Noting that the version, Zenodo archive, and paper all look good.
