### 2020-11-09 #2598 @kthyng
@dpsanders I figured you had discussed it with reviewers but wanted to be sure. Can you check all items that you have heard from reviewers that they are satisfied with? One reviewer is listed twice, which is in turn listed on the paper, so we want all the reviewers to be properly identified.
