### 2020-11-12 #2598 @kthyng
> @kthyng I don't seem to be able to remove the duplicate reviewer for some reason.

I think it looks ok now — you successfully removed the reviewer and I just added back in.
