### 2020-11-12 #2598 @davidjohannesmeyer
> @davidjohannesmeyer Sorry for being pedantic here, but is there is a reason that you didn't check a couple of the boxes?

I forgot to check the code of conduct - thanks.

David
