### 2020-11-12 #2598 @davidjohannesmeyer
> @dpsanders I can see all three in my version — maybe you have a cached version.
>
> I would like to see all the reviewer boxes checked, either by the reviewers themselves or by @dpsanders if you feel clear from your interactions in the review that they should be checked. Then we can proceed. Again sorry for being pedantic but there is really no point to having check boxes at all if we don't fully use them.

Ah OK, performance was unchecked, too.
