### 2020-11-12 #2598 @dpsanders
@kthyng I think we should be ready now. Thanks everybody for polishing off these details.
